# swagger.yaml
openapi: 3.0.3
info:
  title: API REST Défis Sportifs
  version: 1.0.0
  description: |
    Spécification OpenAPI pour l'application Défis Sportifs basée sur le cahier des charges.
servers:
  - url: https://91.168.22.101/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        status: { type: integer }
        error: { type: string }
        message: { type: string }
...
    UserCreate:
      type: object
      required: [nom, prenom, email, password, dateCGU]
      properties:
        nom: { type: string }
        prenom: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password }
        dateCGU: { type: string, format: date-time }
...
    AuthLogin:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    AuthToken:
      type: object
      properties:
        token: { type: string }
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Inscription d'un utilisateur
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '201': { description: Utilisateur créé (email de vérification envoyé) }
        '400': { description: Données invalides, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Connexion et obtention d'un JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthLogin' }
      responses:
        '200':
          description: Jeton renvoyé
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthToken' } } }
        '401': { description: Invalid credentials. }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Demande de réinitialisation du mot de passe (mail envoyé)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '204': { description: Mail envoyé si l'email existe }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Réinitialise le mot de passe via token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, format: password }
      responses:
        '204': { description: Mot de passe mis à jour }
...